refactor: split editor and universe components
