+++ src/editor/index.ts
@@ -1,0 +1,3 @@
+export { default as Header } from './Header';
+export { default as Sidebar } from './Sidebar';
+export { default as CharacterForm } from './CharacterForm';
