+++ src/universe/index.ts
@@ -1,0 +1,2 @@
+export { default as CharacterForm } from './CharacterForm';
+export { default as LocationForm } from './LocationForm';
