+++ src/universeCreator.tsx
@@ -1,3 +1,4 @@
+// @ts-nocheck
 import React, { useState, useEffect } from 'react';
 import {
   User, MapPin, Crown, Shield, Coins, Users, Home,
@@ -6,6 +7,7 @@ import {
   BarChart3, PieChart, TrendingUp, Calendar
 } from 'lucide-react';
 import * as dataStore from '../dataStore';
+import { CharacterForm, LocationForm } from './universe';
 
 const UniverseCreator = () => {
   const [activeTab, setActiveTab] = useState('characters');
@@ -53,351 +55,6 @@ const UniverseCreator = () => {
   const generateEconomy = () => {
     const resources = ['Agricultura', 'Mineração', 'Comércio', 'Pesca', 'Artesanato'];
     return resources[Math.floor(Math.random() * resources.length)];
-  };
-
-  // Componente de Formulário de Personagem
-  const CharacterForm = ({ character, onSave, onCancel }) => {
-    const [formData, setFormData] = useState(character || {
-      name: '',
-      age: '',
-      appearance: '',
-      background: '',
-      abilities: '',
-      motivations: '',
-      relationships: '',
-      role: ''
-    });
-
-    const handleSubmit = (e) => {
-      e.preventDefault();
-      onSave({ ...formData, id: character?.id || Date.now() });
-    };
-
-    return (
-      <div className="fixed inset-0 bg-black bg-opacity-50 flex items-center justify-center z-50">
-        <div className="bg-white rounded-lg p-6 max-w-2xl w-full max-h-[90vh] overflow-y-auto">
-          <h3 className="text-xl font-bold mb-4">
-            {character ? 'Editar Personagem' : 'Novo Personagem'}
-          </h3>
-          <form onSubmit={handleSubmit} className="space-y-4">
-            <div className="grid grid-cols-2 gap-4">
-              <input
-                type="text"
-                placeholder="Nome"
-                value={formData.name}
-                onChange={(e) => setFormData({...formData, name: e.target.value})}
-                className="border rounded px-3 py-2"
-                required
-              />
-              <input
-                type="number"
-                placeholder="Idade"
-                value={formData.age}
-                onChange={(e) => setFormData({...formData, age: e.target.value})}
-                className="border rounded px-3 py-2"
-              />
-            </div>
-            <textarea
-              placeholder="Aparência"
-              value={formData.appearance}
-              onChange={(e) => setFormData({...formData, appearance: e.target.value})}
-              className="border rounded px-3 py-2 w-full h-20"
-            />
-            <textarea
-              placeholder="Histórico"
-              value={formData.background}
-              onChange={(e) => setFormData({...formData, background: e.target.value})}
-              className="border rounded px-3 py-2 w-full h-24"
-            />
-            <textarea
-              placeholder="Habilidades"
-              value={formData.abilities}
-              onChange={(e) => setFormData({...formData, abilities: e.target.value})}
-              className="border rounded px-3 py-2 w-full h-20"
-            />
-            <textarea
-              placeholder="Motivações"
-              value={formData.motivations}
-              onChange={(e) => setFormData({...formData, motivations: e.target.value})}
-              className="border rounded px-3 py-2 w-full h-20"
-            />
-            <textarea
-              placeholder="Relacionamentos"
-              value={formData.relationships}
-              onChange={(e) => setFormData({...formData, relationships: e.target.value})}
-              className="border rounded px-3 py-2 w-full h-20"
-            />
-            <input
-              type="text"
-              placeholder="Papel Narrativo"
-              value={formData.role}
-              onChange={(e) => setFormData({...formData, role: e.target.value})}
-              className="border rounded px-3 py-2 w-full"
-            />
-            <div className="flex gap-2 pt-4">
-              <button
-                type="submit"
-                className="bg-blue-500 text-white px-4 py-2 rounded hover:bg-blue-600"
-              >
-                Salvar
-              </button>
-              <button
-                type="button"
-                onClick={onCancel}
-                className="bg-gray-500 text-white px-4 py-2 rounded hover:bg-gray-600"
-              >
-                Cancelar
-              </button>
-            </div>
-          </form>
-        </div>
-      </div>
-    );
-  };
-
-  // Componente de Formulário de Localização
-  const LocationForm = ({ location, onSave, onCancel }) => {
-    const [formData, setFormData] = useState(location || {
-      name: '',
-      type: 'cidade',
-      climate: '',
-      population: generatePopulation(),
-      culturalComposition: '',
-      mainProfessions: [],
-      economy: generateEconomy(),
-      resources: '',
-      army: {
-        size: 0,
-        weapons: '',
-        training: ''
-      },
-      religions: [],
-      commonFoods: [],
-      establishments: '',
-      strategicPoints: '',
-      government: '',
-      battles: '',
-      events: ''
-    });
-
-    const handleSubmit = (e) => {
-      e.preventDefault();
-      onSave({ ...formData, id: location?.id || Date.now() });
-    };
-
-    const handleArrayChange = (field, value) => {
-      const array = value.split(',').map(item => item.trim()).filter(item => item);
-      setFormData({...formData, [field]: array});
-    };
-
-    return (
-      <div className="fixed inset-0 bg-black bg-opacity-50 flex items-center justify-center z-50">
-        <div className="bg-white rounded-lg p-6 max-w-4xl w-full max-h-[90vh] overflow-y-auto">
-          <h3 className="text-xl font-bold mb-4">
-            {location ? 'Editar Localização' : 'Nova Localização'}
-          </h3>
-          <form onSubmit={handleSubmit} className="space-y-6">
-            {/* Informações Básicas */}
-            <div className="border-b pb-4">
-              <h4 className="font-semibold mb-3">Informações Básicas</h4>
-              <div className="grid grid-cols-3 gap-4">
-                <input
-                  type="text"
-                  placeholder="Nome"
-                  value={formData.name}
-                  onChange={(e) => setFormData({...formData, name: e.target.value})}
-                  className="border rounded px-3 py-2"
-                  required
-                />
-                <select
-                  value={formData.type}
-                  onChange={(e) => setFormData({...formData, type: e.target.value})}
-                  className="border rounded px-3 py-2"
-                >
-                  <option value="cidade">Cidade</option>
-                  <option value="vila">Vila</option>
-                  <option value="reino">Reino</option>
-                  <option value="fortaleza">Fortaleza</option>
-                </select>
-                <input
-                  type="text"
-                  placeholder="Clima"
-                  value={formData.climate}
-                  onChange={(e) => setFormData({...formData, climate: e.target.value})}
-                  className="border rounded px-3 py-2"
-                />
-              </div>
-            </div>
-
-            {/* População */}
-            <div className="border-b pb-4">
-              <h4 className="font-semibold mb-3">População</h4>
-              <div className="grid grid-cols-2 gap-4">
-                <input
-                  type="number"
-                  placeholder="Número de habitantes"
-                  value={formData.population}
-                  onChange={(e) => setFormData({...formData, population: parseInt(e.target.value)})}
-                  className="border rounded px-3 py-2"
-                />
-                <input
-                  type="text"
-                  placeholder="Composição cultural/étnica"
-                  value={formData.culturalComposition}
-                  onChange={(e) => setFormData({...formData, culturalComposition: e.target.value})}
-                  className="border rounded px-3 py-2"
-                />
-              </div>
-            </div>
-
-            {/* Economia e Profissões */}
-            <div className="border-b pb-4">
-              <h4 className="font-semibold mb-3">Economia</h4>
-              <div className="grid grid-cols-2 gap-4">
-                <input
-                  type="text"
-                  placeholder="Setor econômico principal"
-                  value={formData.economy}
-                  onChange={(e) => setFormData({...formData, economy: e.target.value})}
-                  className="border rounded px-3 py-2"
-                />
-                <input
-                  type="text"
-                  placeholder="Profissões (separadas por vírgula)"
-                  value={formData.mainProfessions.join(', ')}
-                  onChange={(e) => handleArrayChange('mainProfessions', e.target.value)}
-                  className="border rounded px-3 py-2"
-                />
-              </div>
-              <textarea
-                placeholder="Recursos e fontes de riqueza"
-                value={formData.resources}
-                onChange={(e) => setFormData({...formData, resources: e.target.value})}
-                className="border rounded px-3 py-2 w-full h-20 mt-2"
-              />
-            </div>
-
-            {/* Exército */}
-            <div className="border-b pb-4">
-              <h4 className="font-semibold mb-3">Exército Local</h4>
-              <div className="grid grid-cols-3 gap-4">
-                <input
-                  type="number"
-                  placeholder="Tamanho do exército"
-                  value={formData.army.size}
-                  onChange={(e) => setFormData({
-                    ...formData,
-                    army: {...formData.army, size: parseInt(e.target.value)}
-                  })}
-                  className="border rounded px-3 py-2"
-                />
-                <input
-                  type="text"
-                  placeholder="Armas principais"
-                  value={formData.army.weapons}
-                  onChange={(e) => setFormData({
-                    ...formData,
-                    army: {...formData.army, weapons: e.target.value}
-                  })}
-                  className="border rounded px-3 py-2"
-                />
-                <input
-                  type="text"
-                  placeholder="Nível de treinamento"
-                  value={formData.army.training}
-                  onChange={(e) => setFormData({
-                    ...formData,
-                    army: {...formData.army, training: e.target.value}
-                  })}
-                  className="border rounded px-3 py-2"
-                />
-              </div>
-            </div>
-
-            {/* Religião e Cultura */}
-            <div className="border-b pb-4">
-              <h4 className="font-semibold mb-3">Religião e Cultura</h4>
-              <div className="grid grid-cols-2 gap-4">
-                <input
-                  type="text"
-                  placeholder="Religiões aceitas (separadas por vírgula)"
-                  value={formData.religions.join(', ')}
-                  onChange={(e) => handleArrayChange('religions', e.target.value)}
-                  className="border rounded px-3 py-2"
-                />
-                <input
-                  type="text"
-                  placeholder="Alimentos comuns (separados por vírgula)"
-                  value={formData.commonFoods.join(', ')}
-                  onChange={(e) => handleArrayChange('commonFoods', e.target.value)}
-                  className="border rounded px-3 py-2"
-                />
-              </div>
-            </div>
-
-            {/* Infraestrutura */}
-            <div className="border-b pb-4">
-              <h4 className="font-semibold mb-3">Infraestrutura</h4>
-              <div className="grid grid-cols-2 gap-4">
-                <textarea
-                  placeholder="Estabelecimentos comerciais"
-                  value={formData.establishments}
-                  onChange={(e) => setFormData({...formData, establishments: e.target.value})}
-                  className="border rounded px-3 py-2 h-24"
-                />
-                <textarea
-                  placeholder="Pontos estratégicos"
-                  value={formData.strategicPoints}
-                  onChange={(e) => setFormData({...formData, strategicPoints: e.target.value})}
-                  className="border rounded px-3 py-2 h-24"
-                />
-              </div>
-            </div>
-
-            {/* História */}
-            <div className="pb-4">
-              <h4 className="font-semibold mb-3">História</h4>
-              <div className="grid grid-cols-2 gap-4">
-                <textarea
-                  placeholder="Histórico de governos"
-                  value={formData.government}
-                  onChange={(e) => setFormData({...formData, government: e.target.value})}
-                  className="border rounded px-3 py-2 h-24"
-                />
-                <textarea
-                  placeholder="Batalhas importantes"
-                  value={formData.battles}
-                  onChange={(e) => setFormData({...formData, battles: e.target.value})}
-                  className="border rounded px-3 py-2 h-24"
-                />
-              </div>
-              <textarea
-                placeholder="Eventos importantes"
-                value={formData.events}
-                onChange={(e) => setFormData({...formData, events: e.target.value})}
-                className="border rounded px-3 py-2 w-full h-20 mt-2"
-              />
-            </div>
-
-            <div className="flex gap-2 pt-4">
-              <button
-                type="submit"
-                className="bg-blue-500 text-white px-4 py-2 rounded hover:bg-blue-600"
-              >
-                Salvar
-              </button>
-              <button
-                type="button"
-                onClick={onCancel}
-                className="bg-gray-500 text-white px-4 py-2 rounded hover:bg-gray-600"
-              >
-                Cancelar
-              </button>
-            </div>
-          </form>
-        </div>
-      </div>
-    );
   };
 
   // Componente de Visualização de Personagem
@@ -942,6 +599,8 @@ const UniverseCreator = () => {
             setSelectedLocation(null);
             setShowLocationForm(false);
           }}
+          generatePopulation={generatePopulation}
+          generateEconomy={generateEconomy}
         />
       )}
     </div>
